// appointza-ui-canvas/src/models/users.model.ts
export class UsersRegisterReq {
  organisationimageid: number = 0;
  organisationname: string = '';
  organisationgstnumber: string = '';
  organisationlogo: number = 0;
  secondarytypecode: string = ""
  secondarytype: number = 0
  primarytype: number = 0
  primarytypecode: string = ""
  organisationtype: number = 0;
  latitude: number = 0;
  longitude: number = 0;
  googlelocation: string = '';
  locationname: string = '';
  locationaddressline1: string = '';
  locationaddressline2: string = '';
  locationcity: string = '';
  locationstate: string = '';
  locationcountry: string = '';
  locationpincode: string = '';
  username: string = '';
  useremail: string = '';
  usermobile: string = '';
  usermobilecountrycode: string = '';
  userdesignation: string = '';
  profileimage: number = 0;
  plan_code: string = '';
  referral_code: string = '';
}

export class UsersRegisterRes {
  mobile: string = '';
  name: string = '';
}

export class UsersLoginReq {
  mobile: string = '';
  otp: string = '';
}

export class UsersContext {
  userid: number = 0;
  usermobile: string = '';
  username: string = '';
  useremail: string = '';
  userimageid: number = 0;
  userpermission: UsersPermissionData = new UsersPermissionData();
  organisationid: number = 0;
  organisationname: string = '';
  organisationlocationid: number = 0;
  organisationlocationname: string = '';
  refreshtoken: string = '';
  accesstoken: string = '';
  static value: any;
}

export class UsersPermissionData {
  // New editandview privileges (single boolean for each)
  editandviewDashboard: boolean = false;
  editandviewAppointments: boolean = false;
  editandviewEvents: boolean = false;
  editandviewCreateService: boolean = false;
  editandviewCreateEvent: boolean = false;
  editandviewClients: boolean = false;
  editandviewBusinessHours: boolean = false;
  editandviewStaffManagement: boolean = false;
  editandviewLocationManagement: boolean = false;
  editandviewTemplates: boolean = false;
  editandviewPaymentSettings: boolean = false;
  editandviewBusinessvalues: boolean = false;
}

export class UsersPermissionGroupData {
  view: boolean = false;
  manage: boolean = false;
}

export class UsersGetOtpReq {
  mobile: string = '';
}